src/third_party/blink/web_tests/svg/custom/rounded-rect-update-to-empty-expected.txt

CONSOLE ERROR: line 9: Error: <rect> attribute width: A negative value is not valid. ("-100")
